Update Copado.yml
Removing a reference (e.g. userPermissions) on a permissionset will set the removed permission to false upon deployment to a target, even if the removed permission is already true in the target.

This native Salesforce MDAPI behavior for permissionset is different, and contrary, to the native Salesforce MDAPI behavior for profile, so it causes confusion for Customers and Partners, especially those newer to the MDAPI or to YAML.

Feedback from both Customers and Partners is that it is best for us to remove any YAML example related to permissionset and avoid misunderstanding or misuse.